fix(RHTAPWATCH-1174): 'cannot set blockOwnerDeletion' error

+++ internal/controller/projectdevelopmentstream_controller_test.go
@@ -24,6 +24,7 @@ import (
 	"math/rand"
 	"os"
 	"path/filepath"
+	"strings"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -54,7 +55,7 @@ var _ = Describe("ProjectDevelopmentStream Controller", func() {
 				testNsN = types.NamespacedName{Namespace: testNs, Name: pdsName}
 
 				for _, resFile := range resFiles {
-					applyFile(ctx, k8sClient, resFile, testNs)
+					applySampleFile(ctx, k8sClient, resFile, testNs)
 				}
 			})
 
@@ -62,8 +63,8 @@ var _ = Describe("ProjectDevelopmentStream Controller", func() {
 				var err error
 
 				controllerReconciler := &ProjectDevelopmentStreamReconciler{
-					Client: k8sClient,
-					Scheme: k8sClient.Scheme(),
+					Client: saClient,
+					Scheme: saClient.Scheme(),
 				}
 
 				By("Setting the owner reference")
@@ -113,18 +114,12 @@ var _ = Describe("ProjectDevelopmentStream Controller", func() {
 	)
 })
 
-func resourceFromFile(fname string, resource client.Object) {
-	testhelpers.ResourceFromFile(
+func applySampleFile(ctx context.Context, k8sClient client.Client, fname string, ns string) {
+	testhelpers.ApplyFile(
+		ctx, k8sClient,
 		filepath.Join("..", "..", "config", "samples", fname),
-		resource,
+		ns,
 	)
-}
-
-func applyFile(ctx context.Context, k8sClient client.Client, fname string, ns string) {
-	var res unstructured.Unstructured
-	resourceFromFile(fname, &res)
-	res.SetNamespace(ns)
-	Expect(k8sClient.Create(ctx, &res)).To(Succeed())
 }
 
 func setupTestNamespace(ctx context.Context, k8sClient client.Client) string {
@@ -140,14 +135,18 @@ func setupTestNamespace(ctx context.Context, k8sClient client.Client) string {
 			break
 		}
 		Expect(err).NotTo(HaveOccurred())
-		Expect(k8sClient.Delete(ctx, &ns)).To(Succeed())
+		if !keepNamespaces() {
+			Expect(k8sClient.Delete(ctx, &ns)).To(Succeed())
+		}
 		// Add a random number to the name to make a unique NS name so we don't
 		// have to wait for the deletion to finish
 		nsName = fmt.Sprintf("test-ns-%d-%d", GinkgoParallelProcess(), rand.Intn(10000))
 	}
 	ns = corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: nsName}}
 	Expect(k8sClient.Create(ctx, &ns)).To(Succeed())
-	DeferCleanup(k8sClient.Delete, &ns)
+	if !keepNamespaces() {
+		DeferCleanup(k8sClient.Delete, &ns)
+	}
 	return nsName
 }
 
@@ -199,4 +198,8 @@ func dropUncomparableMetadata(obj *unstructured.Unstructured) {
 		}
 	}
 	Expect(unstructured.SetNestedField(obj.Object, nmd, "metadata")).To(Succeed())
+}
+
+func keepNamespaces() bool {
+	return strings.ToLower(os.Getenv("KEEP_TEST_NAMESPACES")) == "true"
 }
